Norms expandables implementation
